Post-processor: 1. Redesigned to have common tuning&saving routine. 2. Many tuning plot parameters have been added. 3. Support for more than one plot of the same field. 4. Some minor corrections.
git-svn-id: svn://192.168.1.100/sdvigus/trunk@63 c0ced808-5a4b-11de-9a5a-7fbce8189d0a

+++ src/@VoronoiPlot/compute.m
@@ -97,6 +97,6 @@ parfor iel = 1:num_elem
     [ignore, vcells(iel,:)] = min([part_dp_dist, bpart_dp_dist], [], 2);
     
 end
-vcells = uint8(vcells);
+vcells = uint16(vcells);
 
 end
